refactor: Move dashboard aggregation into the shared services module
Lift the dashboard aggregation (summary, weekly per-task hours and the
bucketed time series, plus the bucket/hours helpers) out of the
`dashboard.py` route and into `services.py`, so the same implementation
backs both the API routes and the upcoming on-demand Telegram report.

The route handlers are now thin wrappers that scope the call to
`request.auth` and translate a domain `ServiceError` into a Ninja
`HttpError`. No behaviour change.
